Refactor CCWWAD
