+++ services/ui-src/src/measures/2021/CCWAD/data.ts
