--- services/ui-src/src/measures/2021/CCWAD/questions/data/index.ts
@@ -1,1 +1,0 @@
-export * as PMD from "./perfromanceMeasureData";
