+++ services/ui-src/src/measures/2021/CCWAD/types.ts
@@ -1,4 +1,4 @@
-import * as Types from "../../CommonQuestions/types";
+import * as Types from "../CommonQuestions/types";
 
 export namespace Measure {
   export interface Props {
